init commit

+++ app/src/main/java/com/tools/sms/adapter/Templatedapter.java
@@ -91,6 +91,6 @@ public class Templatedapter extends RecyclerView.Adapter<Templatedapter.ViewHold
     }
 
     public interface OnItemLongClickListenser {
-        void onItemLongClick(int id,int postion);
+        void onItemLongClick(long id,int postion);
     }
 }
